requirements.txt
numpy>=1.26,<2.0
pandas>=2.0,<3.0
tensorflow==2.16.1
scikit-learn==1.4.0
tqdm
pillow
matplotlib
seaborn
flask
flask-sqlalchemy